services/ingestor/requirements.txt
# Ingestor - Pipeline de procesamiento de documentos
# Python 3.12

# Google Cloud
google-cloud-storage==2.18.2
google-cloud-firestore==2.19.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.27.0
google-cloud-pubsub==2.25.2
google-cloud-logging==3.11.3
google-cloud-vision==3.8.1
google-auth==2.35.0

# PDF processing
pdfminer.six==20231228
pdfplumber==0.11.4
PyPDF2==3.0.1

# Table extraction from PDFs
# camelot-py[cv]==0.11.0  # Commented out due to Python 3.12 compatibility issues
# tabula-py==2.9.3        # Commented out due to Python 3.12 compatibility issues

# Excel/Spreadsheet processing
openpyxl==3.1.5
xlrd==2.0.1
pandas==2.2.3
numpy==2.1.3

# Word documents
python-docx==1.1.2

# OCR
pytesseract==0.3.13
Pillow==11.0.0

# Text processing and NLP
nltk==3.9.1
# spacy==3.8.2  # Commented out due to Python 3.12 compatibility issues

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2

# Utilities
python-dotenv==1.0.1
python-magic==0.4.27
chardet==5.2.0
httpx==0.28.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# FastAPI form handling
python-multipart==0.0.12

# Logging
structlog==24.4.0

